refactor: move CVEs handling to a dedicated service
